+++ test/fixtures/generator-and-minimizer-animation.js
@@ -1,0 +1,2 @@
+require("./animation-test.gif");
+require("./animation-test.gif?as=webp");