+++ types/utils.d.ts
@@ -44,7 +44,7 @@ export type SizeSuffix = (width: number, height: number) => string;
  * @typedef {() => Promise<T>} Task
  */
 /**
- * Run tasks with limited concurency.
+ * Run tasks with limited concurrency.
  * @template T
  * @param {number} limit - Limit of tasks that run at once.
  * @param {Task<T>[]} tasks - List of tasks to run.